requirements.txt
libusb1==3.0.0
